+++ SanKeApp/Phone/UI/主框架/PagedListViewControllerBase.m
@@ -111,7 +111,7 @@
         self.dataFetcher.pageSize = 20;
     }
     @weakify(self);
-    [self.dataFetcher startWithBlock:^(int total, NSArray *retItemArray, NSError *error) {
+    [self.dataFetcher startWithBlock:^(NSInteger total, NSArray *retItemArray, NSError *error) {
         @strongify(self); if (!self) return;
         
         [self tableViewWillRefresh];
@@ -161,7 +161,7 @@
     [self.dataFetcher stop];
     self.dataFetcher.lastID = self.dataFetcher.lastID + self.dataFetcher.pageSize;
     @weakify(self);
-    [self.dataFetcher startWithBlock:^(int total, NSArray *retItemArray, NSError *error) {
+    [self.dataFetcher startWithBlock:^(NSInteger total, NSArray *retItemArray, NSError *error) {
         @strongify(self); if (!self) return;
         [self->_footer endRefreshing];
         if (error) {
